requirements.txt
absl-py==0.7.1
astor==0.7.1
audioread==2.1.8
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
cycler==0.10.0
dask==1.2.2
decorator==4.4.0
dill==0.2.9
distributed==1.28.1
future==0.17.1
gast==0.2.2
google-pasta==0.1.6
googleapis-common-protos==1.5.10
grpcio==1.20.1
h5py==2.9.0
HeapDict==1.0.0
idna==2.8
joblib==0.13.2
kaggle==1.5.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
librosa==0.6.3
llvmlite==0.29.0
Markdown==3.1
matplotlib==3.0.3
mock==3.0.5
msgpack==0.6.1
numba==0.44.0
numpy==1.16.3
pandas==0.23.4
patsy==0.5.1
promise==2.2.1
protobuf==3.7.1
psutil==5.6.2
pyparsing==2.4.0
python-dateutil==2.8.0
python-slugify==3.0.2
pytz==2019.1
pywt==1.0.6
PyYAML==5.1
requests==2.21.0
resampy==0.2.1
scikit-learn==0.21.0
scipy==1.2.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
statsmodels==0.9.0
tb-nightly==1.14.0a20190301
tblib==1.4.0
tensorboard==1.13.1
tensorflow-datasets==1.0.2
tensorflow-estimator==1.13.0
tensorflow-gpu==2.0.0a0
tensorflow-metadata==0.13.0
termcolor==1.1.0
text-unidecode==1.2
tf-estimator-nightly==1.14.0.dev2019030115
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
tsfresh==0.11.2
urllib3==1.24.3
Werkzeug==0.15.2
wrapt==1.11.1
zict==0.1.4
